+++ deepinv/datasets/fastmri.py
@@ -18,7 +18,8 @@ Key modifications:
 """
 
 from pathlib import Path
-from typing import Any, Callable, NamedTuple, Optional, Union, Tuple
+from contextlib import contextmanager
+from typing import Any, Callable, NamedTuple, Optional, Union, Tuple, Dict
 from collections import defaultdict
 import pickle
 import warnings
@@ -27,7 +28,8 @@ import h5py
 from tqdm import tqdm
 import torch
 from torchvision.transforms import Compose, CenterCrop
-from deepinv.datasets.utils import ToComplex, Rescale, download_archive
+
+from deepinv.datasets.utils import ToComplex, Rescale, download_archive, loadmat
 from deepinv.utils.demo import get_image_url
 
 
@@ -235,14 +237,59 @@ class FastMRISliceDataset(torch.utils.data.Dataset):
     """
 
     @staticmethod
-    def torch_shuffle(x: list, generator=None):
+    def torch_shuffle(x: list, generator: torch.Generator = None) -> list:
+        """Shuffle list reproducibly using torch generator.
+
+        :param list x: list to be shuffled
+        :param torch.Generator generator: torch Generator.
+        :return list: shuffled list
+        """
         return [x[i] for i in torch.randperm(len(x), generator=generator).tolist()]
 
-    class SliceSampleFileIdentifier(NamedTuple):
-        """Data structure for identifying specific slices within MRI data files."""
+    class SliceSampleID(NamedTuple):
+        """Data structure containing ID and metadata of specific slices within MRI data files."""
 
         fname: Path
         slice_ind: int
+        metadata: Dict[str, Any]
+
+    @contextmanager
+    def metadata_cache_manager(self, root: Union[str, Path], samples: Any):
+        """Read/write metadata cache file for populating list of sample ids.
+
+        :param Union[str, pathlib.Path] root: root dir to save to metadata cache
+        :param Any samples: iterable (list, dict etc.) for populating with samples to read/write to metadata cache
+        :yield: samples, either populated from metadata cache, or blank, to be yielded to be written to.
+        """
+        if self.load_metadata_from_cache and os.path.exists(self.metadata_cache_file):
+            with open(self.metadata_cache_file, "rb") as f:
+                dataset_cache = pickle.load(f)
+                if dataset_cache.get(root) is None:
+                    raise ValueError(
+                        "`metadata_cache_file` doesn't contain the metadata. Please "
+                        + "either deactivate `load_dataset_from_cache` OR set `metadata_cache_file` properly."
+                    )
+                print(f"Using dataset cache from {self.metadata_cache_file}.")
+                samples = dataset_cache[root]
+
+            yield samples
+
+        else:
+            if self.load_metadata_from_cache and not os.path.exists(
+                self.metadata_cache_file
+            ):
+                warnings.warn(
+                    f"Couldn't find dataset cache at {self.metadata_cache_file}. Loading dataset from scratch."
+                )
+
+            yield samples
+
+            if self.save_metadata_to_cache:
+                dataset_cache = {}
+                dataset_cache[root] = samples
+                print(f"Saving dataset cache to {self.metadata_cache_file}.")
+                with open(self.metadata_cache_file, "wb") as cache_f:
+                    pickle.dump(dataset_cache, cache_f)
 
     def __init__(
         self,
@@ -261,6 +308,9 @@ class FastMRISliceDataset(torch.utils.data.Dataset):
         self.test = test
         self.transform_kspace = transform_kspace
         self.transform_target = transform_target
+        self.load_metadata_from_cache = load_metadata_from_cache
+        self.save_metadata_to_cache = save_metadata_to_cache
+        self.metadata_cache_file = metadata_cache_file
 
         if not os.path.isdir(root):
             raise ValueError(
@@ -278,74 +328,62 @@ class FastMRISliceDataset(torch.utils.data.Dataset):
             raise ValueError('slice_index must be "all", "random", "middle", or int.')
 
         # Load all slices
-        self.sample_identifiers = defaultdict(list)
         all_fnames = sorted(list(Path(root).iterdir()))
 
-        if load_metadata_from_cache and os.path.exists(metadata_cache_file):
-            with open(metadata_cache_file, "rb") as f:
-                dataset_cache = pickle.load(f)
-                if dataset_cache.get(root) is None:
-                    raise ValueError(
-                        "`metadata_cache_file` doesn't contain the metadata. Please"
-                        + "either deactivate `load_dataset_from_cache` OR set `metadata_cache_file` properly."
-                    )
-                print(f"Using dataset cache from {metadata_cache_file}.")
-                self.sample_identifiers = dataset_cache[root]
-        else:
-            if load_metadata_from_cache and not os.path.exists(metadata_cache_file):
-                warnings.warn(
-                    f"Couldn't find dataset cache at {metadata_cache_file}. Loading dataset from scratch."
-                )
-
-            for fname in tqdm(all_fnames):
-                with h5py.File(fname, "r") as hf:
-                    for i in range(hf["kspace"].shape[0]):
-                        self.sample_identifiers[str(fname)].append(
-                            self.SliceSampleFileIdentifier(fname, i)
+        with self.metadata_cache_manager(root, defaultdict(list)) as samples:
+            if len(samples) == 0:
+                for fname in tqdm(all_fnames):
+                    metadata = self._retrieve_metadata(fname)
+                    for slice_ind in range(metadata["num_slices"]):
+                        samples[str(fname)].append(
+                            self.SliceSampleID(fname, slice_ind, metadata)
                         )
 
-            if save_metadata_to_cache:
-                dataset_cache = {}
-                dataset_cache[root] = self.sample_identifiers
-                print(f"Saving dataset cache to {metadata_cache_file}.")
-                with open(metadata_cache_file, "wb") as cache_f:
-                    pickle.dump(dataset_cache, cache_f)
+            self.samples = samples
 
         # Random slice subsampling
         if slice_index != "all":
-            for fname, samples in self.sample_identifiers.items():
+            for fname, samples in self.samples.items():
                 if isinstance(slice_index, int):
                     chosen_sample = samples[slice_index]
                 elif slice_index == "middle":
                     chosen_sample = samples[len(samples) // 2]
                 elif slice_index == "random":
                     chosen_sample = self.torch_shuffle(samples, generator=rng)[0]
-                self.sample_identifiers[fname] = [chosen_sample]
+                self.samples[fname] = [chosen_sample]
 
         # Randomly keep a portion of MRI volumes
         if subsample_volumes < 1.0:
             subsampled_fnames = self.torch_shuffle(
-                list(self.sample_identifiers.keys()), generator=rng
+                list(self.samples.keys()), generator=rng
             )[: round(len(all_fnames) * subsample_volumes)]
-            self.sample_identifiers = {
-                k: self.sample_identifiers[k] for k in subsampled_fnames
-            }
+            self.samples = {k: self.samples[k] for k in subsampled_fnames}
 
         # Flatten to list of samples
-        self.sample_identifiers = [
-            samp for samps in self.sample_identifiers.values() for samp in samps
-        ]
+        self.samples = [samp for samps in self.samples.values() for samp in samps]
+
+    @staticmethod
+    def _retrieve_metadata(fname: Union[str, Path, os.PathLike]) -> Dict[str, Any]:
+        """Open file and retrieve metadata.
+        Metadata includes number of slices in volume.
+
+        :param Union[str, pathlib.Path, os.PathLike] fname: filename to open
+        :return: metadata dict of key-value pairs.
+        """
+        with h5py.File(fname, "r") as hf:
+            metadata = {"num_slices": hf["kspace"].shape[0]}
+        return metadata
 
     def __len__(self) -> int:
-        return len(self.sample_identifiers)
+        return len(self.samples)
 
     def __getitem__(self, idx: int) -> Tuple[Any, Any]:
         r"""Returns the idx-th sample from the dataset, i.e. kspace of shape (2, (N,) H, W) and target of shape (2, H, W)"""
-        fname, dataslice = self.sample_identifiers[idx]
+        fname, slice_ind, metadata = self.samples[idx]
 
         with h5py.File(fname, "r") as hf:
             kspace = torch.from_numpy(
-                hf["kspace"][dataslice]
+                hf["kspace"][slice_ind]
             )  # ((N,) H, W) dtype complex
             kspace = torch.view_as_real(kspace)  # ((N,) H, W, 2)
             kspace = kspace.moveaxis(-1, -3)  # ((N,) 2, H, W)
@@ -361,7 +399,7 @@ class FastMRISliceDataset(torch.utils.data.Dataset):
                     else "reconstruction_rss"
                 )
                 # to shape (1, H, W)
-                target = torch.from_numpy(hf[recons_key][dataslice]).unsqueeze(0)
+                target = torch.from_numpy(hf[recons_key][slice_ind]).unsqueeze(0)
 
                 if self.transform_target is not None:
                     target = self.transform_target(target)